+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 **-.-.- Changes:**
 
 * Fixed interactable outlines randomly not appearing
+* Fixed Grandparent spawn effect disappearing too early
 
 **1.0.1 Changes:**